requirements.txt
# requirements.txt

asttokens==3.0.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
executing==2.2.0
Flask==3.1.0
idna==3.10
ipwhois==1.3.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.0.5
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
networkx==3.4.2
parso==0.8.4
pexpect==4.9.0
phonenumbers==9.0.2
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-whois==0.9.5
pyvis==0.3.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
stem==1.8.2
traitlets==5.14.3
typing_extensions==4.13.0
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3

# -------------------------------------------------------------
# System-level dependencies (must be installed on your system):
#
# 1. Tor:
#    - Ubuntu: sudo apt-get install tor
#    - Other OS: Please refer to https://www.torproject.org/ for installation instructions.
#
# 2. Active Tools (ensure these are installed and available in your system PATH):
#    - Nmap (e.g., sudo apt-get install nmap)
#    - Gobuster (e.g., sudo apt-get install gobuster or download from https://github.com/OJ/gobuster)
#    - Nikto (e.g., sudo apt-get install nikto)
#    - Hydra (e.g., sudo apt-get install hydra)
#    - Medeusa (e.g., follow installation instructions at https://github.com/lanmaster53/medeusa)
#
# These tools are called via subprocess in our active_tools modules.
# -------------------------------------------------------------
